Move `type` to `ty`
