+++ prusti-encoder/src/encoders/ty/kinds/closure.rs
