+++ prusti-encoder/src/encoders/ty/kinds/never.rs
