+++ prusti-encoder/src/encoders/ty/kinds/opaque.rs
